% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAntibodyMap.R
\name{getAntibodyMap}
\alias{getAntibodyMap}
\title{Get antibody map}
\usage{
getAntibodyMap()
}
\value{
a data.frame with the built in antibody map
}
\description{
Get antibody map
}
\concept{
zeptosens
}

