% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeZeptosensData.R
\name{normalizeZeptosensData}
\alias{normalizeZeptosensData}
\title{Normalize Zeptosens Data}
\usage{
normalizeZeptosensData(array1, array2, totalProteinLevels1, totalProteinLevels2,
  controlProbeIndex, antibodyNum)
}
\arguments{
\item{array1}{FIXME}

\item{array2}{FIXME}

\item{totalProteinLevels1}{total protein levels for array 1 (use readZeptosensData format)}

\item{totalProteinLevels2}{total protein levels for array 2 (use readZeptosensData format)}

\item{controlProbeIndex}{row number that that contains the first control probe}

\item{antibodyNum}{number of antibodies on array}
}
\description{
Normalize Zeptosens Data
}
\details{
Normalize data using inter-array method
}
\concept{
zeptosensPkg
}

