% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZeptosensData.R
\name{plotZeptosensData}
\alias{plotZeptosensData}
\title{Plot Zeptosens Data}
\usage{
plotZeptosensData(normArray1, normArray2, antibodies, excludeTreatments,
  ctrlTreatment, ctrlTime, plotColNames, plotRowNames, plotColors, plotDir)
}
\arguments{
\item{normArray1}{normalized data from array 1}

\item{normArray2}{normalized data from array 2}

\item{antibodies}{vector of antibodies}

\item{excludeTreatments}{vector of treatments to be excluded}

\item{ctrlTreatment}{string, treatment label used as control}

\item{ctrlTime}{string, time label used as control}

\item{plotColNames}{vector, array names}

\item{plotRowNames}{vector, treatment time names}

\item{plotColors}{vector, colors for the two arrays}

\item{plotDir}{string, directory to put plots}
}
\description{
Plot Zeptosens Data
}
\concept{
zeptosens
}

