% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readZeptosensExport.R
\name{readZeptosensExport}
\alias{readZeptosensExport}
\title{Read Zeptosens Export File}
\usage{
readZeptosensExport(data, sampleNameEntries = NULL)
}
\arguments{
\item{data}{Zeptosens data read in by readZeptosensXls}

\item{sampleNameEntries}{FIXME, chracter string, perturbation conditions, #cell line etc (sampleNameEntries <- c('sNumber','sample', 'treatment', NA,'time',NA,NA,NA))#}
}
\description{
Read Zeptosens Export File
}
\details{
Format: sNumber_sample_treatment_dose_time_replicate_date_notes
}
\examples{
tmp <- readZeptosensExport(tmp, c("sNumber", NA, NA, "sample", "treatment", "time", "replicate"))
}
\concept{zeptosens}
