% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectresponders.R
\name{selectResponders}
\alias{selectResponders}
\title{Peak selection}
\usage{
selectResponders(arrayData, responseFile, sdctff)
}
\arguments{
\item{arrayData}{FIXME dataframe from readZeptosensExport}

\item{responseFile}{FIXME character, name of the output file}

\item{sdctff}{FIXME standard deviation threshold for selecting responders}
}
\value{
FIXME
}
\description{
Peak selection
}
\concept{
zeptosens
}

