% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixSigma.R
\name{sixSigma}
\alias{sixSigma}
\title{Peak selection w/ 6 sigma of base}
\usage{
sixSigma(arrayData, baseTreatment, nsamp, antibodyNum, sixsigname)
}
\arguments{
\item{arrayData}{FIXME}

\item{baseTreatment}{FIXME}

\item{nsamp}{number of samples}

\item{antibodyNum}{number of antibodies used}

\item{sixsigname}{FIXME}
}
\value{
FIXME
}
\description{
Peak selection w/ 6 sigma of base
}
\concept{
zeptosens
}

