% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSampleNames.R
\name{splitSampleNames}
\alias{splitSampleNames}
\title{Split Sample Names}
\usage{
splitSampleNames(sampleNames, sampleNameEntries)
}
\arguments{
\item{sampleNames}{}
}
\value{
data.frame a data.frame with the split sample names
}
\description{
Split Sample Names
}
\concept{zeptosens}
