% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeSigma.R
\name{threeSigma}
\alias{threeSigma}
\title{Peak selection w/ 6 sigma of base}
\usage{
threeSigma(arrayData, baseTreatment, nsamp, antibodyNum, sixsigname)
}
\arguments{
\item{arrayData}{FIXME dataframe from readZeptosensExport}

\item{baseTreatment}{FIXME Real value, user defined value of base readout from experiments}

\item{nsamp}{number of samples}

\item{antibodyNum}{number of antibodies used}

\item{sixsigname}{FIXME name of the output file}
}
\value{
FIXME
}
\description{
Peak selection w/ 6 sigma of base
}
\concept{
zeptosens
}

