% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalProteinNormalization.R
\name{totalProteinNormalization}
\alias{totalProteinNormalization}
\title{Normalize Zeptosens Data wrt ttl protein level}
\usage{
totalProteinNormalization(array1, totalProteinLevels1, nSamples,
  antibodyNum)
}
\arguments{
\item{array1}{dataframe from readZeptosensExport function}

\item{totalProteinLevels1}{total protein levels for array 1 (use read.xlsx2 format)}

\item{nSamples}{samples/Ab}

\item{antibodyNum}{number of antibodies on array}
}
\description{
Normalize Zeptosens Data wrt ttl protein level
}
\details{
Normalize wrt ttl prt lvl
}
\concept{zeptosens}
