% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unperturbedNormalization.R
\name{unperturbedNormalization}
\alias{unperturbedNormalization}
\title{Normalize Zeptosens Data wrt unperturbed (or any other condition)}
\usage{
unperturbedNormalization(array1, normalFactor, nSamples, antibodyNum)
}
\arguments{
\item{array1}{data frame from readZeptosensExport function}

\item{normalFactor}{(treatment_name)}

\item{nSamples}{samples/Ab}

\item{antibodyNum}{number of antibodies on array}
}
\description{
Normalize Zeptosens Data wrt unperturbed (or any other condition)
}
\details{
Normalize wrt ttl prt lvl
}
\concept{zeptosens}
