% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectResponders.R
\name{selectResponders}
\alias{selectResponders}
\title{Peak selection}
\usage{
selectResponders(arrayData, responseFile, sdctff)
}
\arguments{
\item{arrayData}{FIXME}

\item{responseFile}{FIXME}

\item{sdctff}{FIXME}
}
\value{
FIXME
}
\description{
Peak selection
}
\concept{
zeptosens
}

