% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalProteinNormalization.R
\name{totalProteinNormalization}
\alias{totalProteinNormalization}
\title{Normalize Zeptosens Data wrt ttl protein level}
\usage{
totalProteinNormalization(Array1, totalProteinLevels1, nsamples, antibodyNum)
}
\arguments{
\item{totalProteinLevels1}{total protein levels for array 1 (use readZeptosensData format)}

\item{nsamples}{samples/Ab}

\item{antibodyNum}{number of antibodies on array}

\item{array1}{FIXME}
}
\description{
Normalize Zeptosens Data wrt ttl protein level
}
\details{
Normalize wrt ttl prt lvl
}
\concept{
zeptosens
}

